Share ID display logic
