+++ src/Entities/PropertyName.tsx
@@ -1,7 +1,6 @@
 import React from "react";
-import { Link } from "react-router-dom";
 import { RDFS_LABEL, JSON_LD_TYPE, RDFS_SUB_CLASS_OF, Label } from "./data";
-import { entityLink } from "./navigation";
+import Value from "./Value";
 
 const PropertyName = ({
   property,
@@ -15,33 +14,16 @@ const PropertyName = ({
     return <b>Labels</b>;
   }
   if (property === JSON_LD_TYPE) {
-    return <b>Type</b>;
+    return <b>Types</b>;
   }
   if (property === RDFS_SUB_CLASS_OF) {
     return <b>Sub Class Of</b>;
   }
-  const display = label || propertyToDisplay(property);
   return (
     <b>
-      <Link to={entityLink(property)}>{display}</Link>
+      <Value value={{ "@id": property }} label={label} />
     </b>
   );
 };
 
 export default PropertyName;
-
-function propertyToDisplay(property: string): string {
-  try {
-    const url = new URL(property);
-    if (url.hash) {
-      return url.hash.substr(1);
-    }
-    if (url.pathname.length > 1) {
-      const parts = url.pathname.split("/");
-      return parts[parts.length - 1];
-    }
-    return property;
-  } catch {
-    return property;
-  }
-}
